+++ python/src/lib.rs
@@ -35,6 +35,8 @@ fn dashu(m: &Bound<'_, PyModule>) -> PyResult<()> {
 
     m.add_function(wrap_pyfunction!(utils::auto, m)?)?;
     m.add_function(wrap_pyfunction!(utils::autos, m)?)?;
+    m.add_function(wrap_pyfunction!(utils::get_precision, m)?)?;
+    m.add_function(wrap_pyfunction!(utils::set_precision, m)?)?;
 
     // module-level math functions
     m.add_function(wrap_pyfunction!(math::sin, m)?)?;